requirements.txt
alchemy==20.5
antiorm==1.2.1
anyio==3.6.2
asgiref==3.4.1
bigcommerce==0.20.1
cached-property==1.5.2
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
db==0.1.1
docutils==0.18.1
fastapi==0.83.0
filelock==3.0.12
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.13.0
idna==2.8
immutables==0.19
importlib-metadata==4.8.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
lockfile==0.12.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
psycopg2-binary==2.9.5
pydantic==1.9.2
PyJWT==2.4.0
python-daemon==2.2.4
python-dotenv==0.20.0
requests==2.22.0
sniffio==1.2.0
SQLAlchemy==1.4.45
sqlalchemy-orm==1.2.3
starlette==0.19.1
typing-inspect==0.8.0
typing_extensions==4.1.1
urllib3==1.25.11
uvicorn==0.16.0
Werkzeug==2.0.3
zipp==3.6.0